Table Join, Table Join List, Signature field types moved to the Pro Plugin, also Twig filters also moved to the plugin.
Field Extra tasks like moving images from folder to folder also became a part of the Pro version.
All this is because of the use of base64_encoding which is not advised by JED.

Where clouse is more strict and more secure.

Where filters are URL encoded now, not base64, and are visible in the URL which makes it more Search Engine Friendly.